+++ clang-tools-extra/clang-tidy/bsl/BslTidyModule.cpp
@@ -10,6 +10,7 @@
 #include "../ClangTidyModule.h"
 #include "../ClangTidyModuleRegistry.h"
 #include "FixedWidthInts.h"
+#include "LiteralsAsciiOnlyCheck.h"
 #include "LiteralsNoOctalCheck.h"
 #include "LiteralsUnsignedSuffixCheck.h"
 #include "LiteralsUppercaseSuffixCheck.h"
@@ -23,6 +24,8 @@ public:
   void addCheckFactories(ClangTidyCheckFactories &CheckFactories) override {
     CheckFactories.registerCheck<FixedWidthInts>(
         "bsl-fixed-width-ints");
+    CheckFactories.registerCheck<LiteralsAsciiOnlyCheck>(
+        "bsl-literals-ascii-only");
     CheckFactories.registerCheck<LiteralsNoOctalCheck>(
         "bsl-literals-no-octal");
     CheckFactories.registerCheck<LiteralsUnsignedSuffixCheck>(