+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -80,6 +80,11 @@ New checks
   Finds non-const global variables as described in check I.2 of C++ Core
   Guidelines.
 
+- New :doc:`bsl-literals-ascii-only
+  <clang-tidy/checks/bsl-literals-ascii-only>` check.
+
+  Ensures only ASCII-encoded string literals are used.
+
 - New :doc:`bsl-literals-no-octal
   <clang-tidy/checks/bsl-literals-no-octal>` check.